# backend/openapi/realtime-service.yaml
# Generated by scripts/split_openapi.py — do not edit.
# Source: openapi/combined.yaml
openapi: 3.1.0
info:
  title: RS3 Market — realtime-service
  version: 2.0.0
  description: 'Endpoints served by **realtime-service**, extracted from the combined specification.


    This copy exists for debugging one service in isolation. Generate clients from `openapi/combined.yaml`,
    which describes the whole gateway surface.


    Generated by `scripts/split_openapi.py` — do not edit by hand.'
servers:
- url: /
  description: realtime-service
security: []
paths:
  /health:
    get:
      tags:
      - health
      operationId: getHealth
      summary: Gateway liveness
      responses:
        '200':
          description: Gateway is up.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
  /auth/register:
    post:
      tags:
      - auth
      operationId: register
      summary: Create a chat account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Account created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredUser'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags:
      - auth
      operationId: login
      summary: Exchange credentials for a JWT
      description: 'The token authenticates the WebSocket at `/ws`, where it is

        passed as the `token` query parameter.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Signed token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /chat/history:
    get:
      tags:
      - chat
      operationId: getChatHistory
      summary: Recent chat messages
      parameters:
      - name: before
        in: query
        description: RFC3339 cursor; returns messages strictly older than this.
        schema:
          type: string
          format: date-time
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          default: 50
      responses:
        '200':
          description: Messages in chronological order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatHistoryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /ws:
    get:
      tags:
      - chat
      operationId: openWebSocket
      summary: Realtime price and chat socket
      description: 'Upgrades to a WebSocket. Not usable through a generated HTTP

        client — connect with the browser''s `WebSocket` API. It is

        documented here so the message payloads are generated as types.


        Send `WsCommand` frames; receive `WsEnvelope` frames whose

        `payload` shape is determined by `type`:


        | `type`      | `payload`        |

        |-------------|------------------|

        | `price`     | `PriceSnapshot`  |

        | `chat`      | `ChatMessage`    |

        | `heartbeat` | `WsHeartbeat`    |

        | `error`     | `WsError`        |


        Price frames arrive only for items subscribed to on this

        connection; chat and heartbeat frames go to every client.

        '
      parameters:
      - name: token
        in: query
        required: true
        description: JWT from `/auth/login`.
        schema:
          type: string
      responses:
        '101':
          description: Switching protocols.
        '401':
          description: Missing or invalid token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
tags:
- name: auth
  description: Account registration and login for chat.
- name: chat
  description: Chat history and the realtime socket.
- name: health
  description: Liveness and upstream status.
components:
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    ChatHistoryResponse:
      type: object
      required:
      - messages
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
    ChatMessage:
      type: object
      required:
      - id
      - user_id
      - username
      - body
      - created_at
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        username:
          type: string
        body:
          type: string
          maxLength: 500
        created_at:
          type: string
          format: date-time
    Credentials:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          format: password
          minLength: 6
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable failure reason.
    HealthStatus:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
    LoginResponse:
      type: object
      required:
      - token
      properties:
        token:
          type: string
          description: JWT; pass as `?token=` when opening `/ws`.
    RegisteredUser:
      type: object
      required:
      - id
      - username
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string